Extract AopCodeGen method signature
